Move app code into lib.rs
